korpexport.formatter: Attach match marker to token
korpexport.formatter: Support attaching match open, match close and match
marker to tokens (the first, the last and any token of a match,
respectively). The token formats may now contain references to the keys
match_open, match_close and match_marker, which are the values of the
corresponding options when appropriate and the empty string for other
tokens. This makes it easier to add match markers to tokens. The default
values for token formats now contain match_open and match_close.

korpexport/formatter.py:
  KorpExportFormatter:
    _option_defaults: token_format, token_noattrs_format,
      token_attr_format: Added {match_open} and {match_close} around the
      actual token information.
    _format_sentence(): Pass match info to _format_tokens for the lists of
      all tokens and match tokens.
    _format_token(): Add format parameters match_open, match_close,
      match_marker when appropriate (the first, the last and any token of a
      match).