korpexport.formatter: Support ?item in corpus info
korpexport.formatter: Support the notation ?item for corpus information
items, such as licence and metadata information (in addition to ?aligned):
the item is included only if any of the corpora in the result has a value
for it. The corpus info items can be used in many of the list-valued
options, such as corpus_info_fields and sentence_fields.

korpexport/formatter.py:
  Support notation ?item for corpus info items in addition to ?aligned.
    (KorpExportFormatter._make_opt_lists)

korpexport/queryresult.py:
  Added get_occurring_corpus_info.